backend/requirements.txt
fastapi
uvicorn[standard]
redis
pydantic
pydantic-settings
python-arango
sse-starlette
python-dotenv
